backend/requirements.txt
asgiref
Django
django-coreapi
django-cors-headers
djangorestframework
django-filter
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
tzdata
Pillow
pandas
setuptools>=58.0.0
drf-yasg==1.21.5
flask
openpyxl==3.1.2
gunicorn